refactor: detailed product page refactoring / memoisation
